refactor: Move env utils into lib/envutils
