+++ lib/envutils/env.go
@@ -1,4 +1,4 @@
-package utils
+package envutils
 
 import (
 	"fmt"
