+++ lib/envutils/env_test.go
@@ -1,4 +1,4 @@
-package utils
+package envutils
 
 import (
 	"github.com/stretchr/testify/assert"
